+++ Gemfile
@@ -1,6 +1,6 @@
 source "https://rubygems.org"
 
-# Specify your gem's dependencies in actionpush.gemspec.
+# Specify your gem's dependencies in action_native_push.gemspec.
 gemspec
 
 gem "sqlite3"
