+++ lib/action_native_push/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module ActionNativePush
+  VERSION = "0.1.0"
+end
