--- lib/action_push/version.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module ActionPush
-  VERSION = "0.1.0"
-end
