+++ lib/generators/action_native_push/install/templates/app/jobs/application_push_notification_job.rb.tt
