+++ lib/generators/action_native_push/install/templates/app/models/application_push_device.rb.tt
@@ -1,4 +1,4 @@
 class ApplicationPushDevice < ActionNativePush::Device
   # Customize TokenError handling (default: destroy!)
-  rescue_from (ActionNativePush::TokenError) { }
+  # rescue_from (ActionNativePush::TokenError) { Rails.logger.error("Device #{id} token is invalid") }
 end
