+++ lib/generators/action_native_push/install/templates/app/models/application_push_notification.rb.tt
@@ -1,9 +1,9 @@
 class ApplicationPushNotification < ActionNativePush::Notification
   # Set a custom job queue_name
-  queue_as :realtime
+  # queue_as :realtime
 
   # Controls whether push notifications are enabled (default: !Rails.env.test?)
-  self.enabled = true
+  # self.enabled = Rails.env.production?
 
   # Define a custom callback to modify or abort the notification before it is sent
   # before_delivery do |notification|
