+++ lib/generators/action_native_push/install/templates/config/push.yml.tt
