--- lib/generators/action_push/install/templates/app/jobs/application_push_notification_job.rb.tt
@@ -1,7 +1,0 @@
-class ApplicationPushNotificationJob < ActionPush::NotificationJob
-  # Enable logging job arguments (default: false)
-  # self.log_arguments = true
-
-  # Report job retries via the `Rails.error` reporter (default: false)
-  # self.report_job_retries = true
-end
