--- lib/generators/action_push/install/templates/app/models/application_push_device.rb.tt
@@ -1,4 +1,0 @@
-class ApplicationPushDevice < ActionPush::Device
-  # Customize TokenError handling (default: destroy!)
-  # rescue_from (ActionPush::TokenError) { Rails.logger.error("Device #{id} token is invalid") }
-end
