+++ test/fixtures/action_native_push/devices.yml
