+++ test/models/application_push_device_test.rb
@@ -2,9 +2,9 @@ require "test_helper"
 
 class ApplicationPushDeviceTest < ActiveSupport::TestCase
   test "TokenErrors are ignored" do
-    notification = ActionPush::Notification.new(title: "Hi!")
+    notification = ActionNativePush::Notification.new(title: "Hi!")
     iphone = application_push_devices(:iphone_6)
-    ActionPush::Service::Apns.any_instance.expects(:push).raises(ActionPush::TokenError)
+    ActionNativePush::Service::Apns.any_instance.expects(:push).raises(ActionNativePush::TokenError)
 
     assert_no_difference -> { ApplicationPushDevice.count } do
       iphone.push(notification)
